Revert "Fixes #30947: Add team quick info popover on entity owner chips (#31108)"
This reverts commit 23a939ecb260a39f416ec2caca4d48c2c57dfb1b.
